refactor(nsq): seperate marshaller and compressot
