+++ transport/nsq/handler/handler.go
@@ -3,8 +3,7 @@ package handler
 import (
 	"context"
 
-	"github.com/alexfalkowski/go-service/compressor"
-	"github.com/alexfalkowski/go-service/marshaller"
+	"github.com/alexfalkowski/go-service/transport/nsq/marshaller"
 	"github.com/alexfalkowski/go-service/transport/nsq/message"
 	"github.com/nsqio/go-nsq"
 )
@@ -17,18 +16,16 @@ type Handler interface {
 // Params for handler.
 type Params struct {
 	Handler    Handler
-	Compressor compressor.Compressor
 	Marshaller marshaller.Marshaller
 }
 
 // New handler for NSQ.
 // nolint:ireturn
 func New(params Params) nsq.Handler {
-	return &handler{Compressor: params.Compressor, Marshaller: params.Marshaller, Handler: params.Handler}
+	return &handler{Marshaller: params.Marshaller, Handler: params.Handler}
 }
 
 type handler struct {
-	Compressor compressor.Compressor
 	Marshaller marshaller.Marshaller
 	Handler
 }
@@ -39,13 +36,7 @@ func (h *handler) HandleMessage(m *nsq.Message) error {
 	}
 
 	var msg message.Message
-
-	bytes, err := h.Compressor.Decompress(m.Body)
-	if err != nil {
-		return err
-	}
-
-	if err := h.Marshaller.Unmarshal(bytes, &msg); err != nil {
+	if err := h.Marshaller.Unmarshal(m.Body, &msg); err != nil {
 		return err
 	}
