+++ transport/nsq/marshaller/msgpack.go
@@ -1,0 +1,11 @@
+package marshaller
+
+import (
+	"github.com/alexfalkowski/go-service/marshaller"
+)
+
+// NewMsgPack for NSQ.
+// nolint:ireturn
+func NewMsgPack() Marshaller {
+	return marshaller.NewMsgPack()
+}
